Jax working, but slow

+++ deluca/lung/controllers/_deep.py
@@ -118,13 +118,13 @@ class Deep(Controller):
       trajectory = jnp.expand_dims(next_errs[-self.history_len:], axis=(0, 1))
       u_in = self.model.apply({"params": self.params},
                               (trajectory @ self.featurizer))
-      return u_in.squeeze().astype(jnp.float64)
+      return u_in.squeeze().astype(jnp.float32)
 
     # changed decay compare from None to float(inf) due to cond requirements
     u_in = jax.lax.cond(
         jnp.isinf(decay), true_func, lambda x: jnp.array(decay), None)
 
-    u_in = jax.lax.clamp(0.0, u_in.astype(jnp.float64), self.clip).squeeze()
+    u_in = jax.lax.clamp(0.0, u_in.astype(jnp.float32), self.clip).squeeze()
 
     # update controller_state
     new_dt = jnp.max(
